:white_check_mark: Restructured project and removed unnecessary logic & bugs
