--- facebook-policy.md
@@ -1,3 +1,0 @@
-# Privacy policy
-
-This app is only for development purposes. The application does not collect any personal information from the user.
