+++ package.json
@@ -4,26 +4,19 @@
   "private": true,
   "dependencies": {
     "@emotion/styled": "^11.11.5",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
     "axios": "^1.6.8",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "jwt-decode": "^4.0.0",
     "qs": "^6.12.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
     "react-social-login-buttons": "^4.1.0",
     "reactjs-social-login": "^2.6.3",
-    "uuid": "^9.0.1",
-    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "start-server": "node server.js",
+    "startFrontend": "react-scripts start",
+    "startBackend": "node src/backend/server.js",
     "start-all": "npm-run-all --parallel start start-server",
     "build": "react-scripts build",
     "test": "react-scripts test",
