+++ src/frontend/App.js
@@ -1,0 +1,16 @@
+import "./styles/App.css";
+import FacebookLogin from "./components/FacebookLogin";
+import GoogleLogin from "./components/GoogleLogin";
+import LinkedInLogin from "./components/LinkedInLogin";
+
+function App() {
+  return (
+    <div className="App">
+      <GoogleLogin />
+      <FacebookLogin />
+      <LinkedInLogin />
+    </div>
+  );
+}
+
+export default App;
