+++ src/frontend/styles/App.css
@@ -1,6 +1,5 @@
 .App {
-  height: 1000px;
-  
+  height: 1000px; 
   margin: 0 auto;
   padding: 30px;
   display: grid;
